requirements.txt
# tool dependencies
pylint==2.12.2
numpy>=1.19.0
pillow==9.0.0
minio==7.1.2
pandas>=1.3.5
pytest==7.1.1